+++ src/index.css
@@ -27,19 +27,8 @@
 
 /* ScrollBar */
 ::-webkit-scrollbar {
-  width: 5px;
-  height: 50px;
+  display: none;
 }
-::-webkit-scrollbar-track {
-  background-color: var(--background-color);
-}
-::-webkit-scrollbar-thumb {
-  background: var(--accent-color);
-}
-::-webkit-scrollbar-thumb:hover {
-  background:var(--accent-color-lighter)
-}
-
 
 /* Selection Color */
 ::selection{
